--- apps/ui/src/components/Schema/index.ts
@@ -1,7 +1,0 @@
-export * from "./Field";
-export * from "./FieldName";
-export * from "./SchemaViewer";
-export * from "./Stats";
-export * from "./StatsDetails";
-export * from "./Type";
-export * from "./utils";
