--- apps/ui/src/components/Trace/Editor/index.ts
@@ -1,6 +1,0 @@
-export * from "./json-viewer";
-export * from "./query-viewer";
-export * from "./utils";
-export * from "./Query";
-export * from "./Editor";
-export * from "./Variables";
