--- apps/ui/src/components/Trace/Editor/json-viewer/index.ts
@@ -1,2 +1,0 @@
-export * from "./JsonValue";
-export * from "./JsonViewer";
