+++ apps/ui/src/components/Trace/Editor/query/argument.tsx
