+++ apps/ui/src/components/Trace/Editor/query/type.tsx
