--- apps/ui/src/components/Trace/TraceStats/index.ts
@@ -1,5 +1,0 @@
-export * from "./SchemaTraces";
-export * from "./Span";
-export * from "./TraceHeader";
-export * from "./TraceViewer";
-export * from "./utils";
