--- apps/ui/src/components/Trace/index.ts
@@ -1,3 +1,0 @@
-export * from "./Editor";
-export * from "./TraceStats";
-export * from "./Trace";
