+++ apps/ui/src/components/Trace/stats/trace-header.tsx
