+++ apps/ui/src/components/Trace/stats/trace-viewer.tsx
@@ -5,13 +5,13 @@ import { useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
 
 import { client } from "../../../client";
-import { Spinner } from "../../../components/utils/Spinner";
 import { Modal } from "../../../context/modal";
 import { IDS } from "../../../testing";
 import { DEFAULT_SLEEP_TIME, sleep } from "../../../utils/sleep";
-import { ModalWindow } from "../../modal/ModalWindow";
-import { OpenModal } from "../../modal/OpenModal";
-import { Span } from "./Span";
+import { OpenModal } from "../../modal/open";
+import { ModalWindow } from "../../modal/window";
+import { Spinner } from "../../utils/Spinner";
+import { Span } from "./span";
 import { createTreeData } from "./utils";
 
 const TraceView = ({ spans }: { spans: TSpan[] }) => {
