+++ apps/ui/src/components/Trace/stats/utils.ts
