--- apps/ui/src/components/index.ts
@@ -1,2 +1,0 @@
-export * from "./Trace";
-export * from "./Schema";
