+++ apps/ui/src/components/info/getting-started.tsx
