+++ apps/ui/src/components/info/how-it-works.tsx
