+++ apps/ui/src/components/info/info-logo.tsx
