+++ apps/ui/src/index.tsx
@@ -5,7 +5,7 @@ import "@fontsource/roboto/700.css";
 import { createRoot } from "react-dom/client";
 import "tailwindcss/tailwind.css";
 
-import { App } from "./App";
+import { App } from "./app";
 import "./index.css";
 
 const root = createRoot(document.getElementById("root") as HTMLElement);
